Added SabreTokenService to obtain Bearer token
Sabre necessiates to obtain a Authorization token, which
usually expires in 7 days from the time of acquisition. This token
needs to be set in every call to a Sabre service via Header parameters
(check Sabre Docs for more details). The implementation is done in a
thread safe manner necessiating the first thread to obtain a lock, while
rest of the threads could await for a signal. Refer to the source code for
more details.